Allow to use the library with ESP8266
It seems that the ESP8266 I2C implementation (which is a software
implementation from what I am aware of) causes communication issues
with the SCD30. It seems not to be related to clock stretching since
this is supported by the software implementation since a while.
In the drawings of the interface description there seems to be a pause
between sending a command and reading back the result. It is not
specified, but by expermentation I found 1ms to be long enough. Use
write_to/read_from separately instead of realying on readfrom_mem to
support a pause between sending the command and reading back values.